Install binaries in local appdata
The launcher currently tries to put the binaries in a subfolder
of the user’s game files, which causes problems with OneDrive
and isn’t really a good place to store these to begin with.

This makes the code uses %LOCALAPPDATA% instead. This seems to be
the best place, as the binaries can be machine-specific, which is
why I didn’t pick %APPDATA%. (This is only relevant to roaming
profiles, but we might as well do it right while we’re at it.)

On Linux, this puts the files in ~/.local/share/Open{RCT2,Loco}